LayoutTests/platform/glib/imported/w3c/web-platform-tests/encrypted-media/clearkey-mp4-update-disallowed-input.https-expected.txt


FAIL update() with invalid response (longer than 64Kb characters) should fail. assert_equals: expected "TypeError" but got "Error"
